some changes with name and content of client
